requirements.txt
python-dotenv==1.0.0
slack-bolt==1.18.0
jira==3.5.2
requests==2.31.0
pandas==2.1.1
numpy==1.26.0
transformers==4.34.0
torch==2.4.1
openai==0.28.1
chromadb==0.4.13
cryptography==41.0.4
python-dateutil==2.8.2
Flask==2.3.3
scipy==1.11.3
aiohttp==3.10.9

# Additional dependencies
pytest==7.4.2
pydantic==2.4.2
tqdm==4.66.1
scikit-learn==1.3.1
matplotlib==3.8.0
seaborn==0.13.0

# Add setuptools to ensure distutils is available
setuptools==68.2.2

# New dependencies for admin interface
Flask-Bcrypt==1.0.1
Flask-WTF==1.1.1
email-validator==2.0.0
flask-paginate==2023.10.8